Set up tests for StatewideTest class
