+++ test/test_helper.rb
@@ -7,3 +7,4 @@ require './test/enrollment_test'
 require './test/enrollment_repository_test'
 require './test/headcount_analyst_test'
 require './test/statewide_test_repository_test'
+require './test/statewide_test_test'
